pyproject.toml
[project]
name = "zola2astro"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.42.0",
    "pyyaml>=6.0.2",
    "toml>=0.10.2",
]
